requirements.txt
huggingface_hub==0.22.2
gradio
huggingface_hub
PyMuPDF
sentence-transformers
numpy
faiss-cpu